backend: fix createPet endpoint
